refactor: move interactor to domain
