+++ src/get-saml-metadata/src/domain/interactors/GetExternalDataRequestModel.ts
