+++ src/get-saml-metadata/src/domain/interactors/GetExternalDataResponseModel.ts
