+++ src/get-saml-metadata/src/domain/interactors/IGetExternalDataInputBoundary.ts
