+++ src/get-saml-metadata/src/domain/interactors/IGetExternalDataOutputBoundary.ts
